+++ src/tests/test_geometry.cpp
@@ -28,10 +28,10 @@ void GeometryTest::render(Renderer* r) {
     r->setBlending(false);
     r->useShader(Renderer::ShaderType::Scene3D);
 
-    float aspect = static_cast<float>(vw_) / vh_;
+    float aspect = static_cast<float>(vw_) / static_cast<float>(vh_);
     r->setProjection(Mat4::perspective(60.0f, aspect, 0.1f, 500.0f));
 
-    float dist = grid_size_ * 2.0f;
+    float dist = static_cast<float>(grid_size_) * 2.0f;
     r->setView(Mat4::lookAt(
         Vec3(dist * 0.7f, dist * 0.5f, dist * 0.7f),
         Vec3(0, 0, 0),
