+++ src/tests/test_instanced_draw.cpp
@@ -104,7 +104,7 @@ void InstancedDrawTest::renderGL3(Renderer& r, GL3Features& gl3) {
     if (shader_ != INVALID_SHADER) {
         r.useCustomShader(shader_);
 
-        float aspect = static_cast<float>(vw_) / vh_;
+        float aspect = static_cast<float>(vw_) / static_cast<float>(vh_);
         Mat4 proj = Mat4::perspective(60.0f, aspect, 0.1f, 5000.0f);
         Mat4 view = Mat4::lookAt(Vec3(0, 80, 120), Vec3(0, 0, 0), Vec3(0, 1, 0));
 
@@ -114,7 +114,7 @@ void InstancedDrawTest::renderGL3(Renderer& r, GL3Features& gl3) {
     } else {
         // Fallback: use built-in shader
         r.useShader(Renderer::ShaderType::Scene3D);
-        float aspect = static_cast<float>(vw_) / vh_;
+        float aspect = static_cast<float>(vw_) / static_cast<float>(vh_);
         r.setProjection(Mat4::perspective(60.0f, aspect, 0.1f, 5000.0f));
         r.setView(Mat4::lookAt(Vec3(0, 80, 120), Vec3(0, 0, 0), Vec3(0, 1, 0)));
         r.setColor(0.6f, 0.7f, 0.8f, 1.0f);
